STORM-974 introduce DefaultEsTupleMapper and reflect codes
