+++ external/storm-elasticsearch/src/test/java/org/apache/storm/elasticsearch/common/EsTestUtil.java
@@ -48,27 +48,7 @@ public class EsTestUtil {
     }
 
     public static EsTupleMapper generateDefaultTupleMapper() {
-        return new EsTupleMapper() {
-            @Override
-            public String getSource(ITuple tuple) {
-                return tuple.getStringByField("source");
-            }
-
-            @Override
-            public String getIndex(ITuple tuple) {
-                return tuple.getStringByField("index");
-            }
-
-            @Override
-            public String getType(ITuple tuple) {
-                return tuple.getStringByField("type");
-            }
-
-            @Override
-            public String getId(ITuple tuple) {
-                return tuple.getStringByField("id");
-            }
-        };
+        return new DefaultEsTupleMapper();
     }
 
     public static Node startEsNode(){
